requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
email-validator
pandas
openpyxl
python-dotenv